src/main/java/net/lesson/webfluxsecurity/security/AuthManager.java
package net.lesson.webfluxsecurity.security;

import lombok.RequiredArgsConstructor;
import net.lesson.webfluxsecurity.domain.UserEntity;
import net.lesson.webfluxsecurity.exception.UnauthorizedException;
import net.lesson.webfluxsecurity.service.UserService;
import org.springframework.security.authentication.ReactiveAuthenticationManager;
import org.springframework.security.core.Authentication;
import org.springframework.stereotype.Component;
import reactor.core.publisher.Mono;

@Component
@RequiredArgsConstructor
public class AuthManager implements ReactiveAuthenticationManager {
    private final UserService service;

    @Override
    public Mono<Authentication> authenticate(Authentication authentication) {
        CustomPrincipal principal = (CustomPrincipal) authentication.getPrincipal();
        return service.getUserById(principal.getId())
                .filter(UserEntity::isEnabled)
                .switchIfEmpty(Mono.error(new UnauthorizedException("User disabled")))
                .map(user -> authentication);
    }
}
